requirements.txt
django-environ
django>=5.1.0
djangorestframework>=3.15.2
django-cors-headers>=4.4.0
channels>=4.0.0
channels-redis>=4.2.0
psycopg2-binary>=2.9.0
Pillow>=10.0.0
celery>=5.4.0
django-celery-beat>=2.6.0
redis>=5.0.0
google-auth>=2.34.0
google-auth-oauthlib>=1.2.1
google-api-python-client>=2.156.0
requests>=2.31.0
pyjwt[crypto]>=2.8.0
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-django>=4.8
cryptography
email-validator
whitenoise>=6.7.0
httpx>=0.27.0
djangorestframework-simplejwt>=5.3.0
